Updated Ban system

+++ strategies/discord.js
@@ -3,13 +3,31 @@ var DiscordStrategy = require('passport-discord').Strategy
 const passport = require('passport');
 const mongoose = require('mongoose');
 const DiscordUser = require('../schemas/DiscordUser.js');
-const banned = [];
 const jwt = require('jsonwebtoken');
+
+
+const { QuickDB } = require('quick.db');
+const db = new QuickDB(); // using default driver
 
 var scopes = ['identify', 'email', 'guilds', 'guilds.join'];
 
-passport.serializeUser((user, done) => done(null, user))
-passport.deserializeUser((user, done) => done(null, user))
+passport.serializeUser((user, done) => {
+  if(checkBan(user.discordId)) return done("You are banned", null)
+  done(null, user)
+})
+passport.deserializeUser((user, done) => {
+  if(checkBan(user.discordId)) return done("You are banned", null)
+  done(null, user)
+})
+
+async function checkBan(userId){
+  const banned = await db.get(userId);
+  if(banned){
+    return true
+  } else {
+    return false
+  }
+}
 
 var discordStrat = passport.use(new DiscordStrategy({
     clientID: '1044716282311888909',
@@ -19,6 +37,7 @@ var discordStrat = passport.use(new DiscordStrategy({
 },
 async (accessToken, refreshToken, profile, done) => {
   try {
+    if(checkBan(profile.id)) return done("You are banned", null)
     const discordUser = await DiscordUser.findOne({ discordId: profile.id });
     if(discordUser) {
       await DiscordUser.deleteOne({ discordId: profile.id });
@@ -36,7 +55,7 @@ async (accessToken, refreshToken, profile, done) => {
       console.log("Deleted and created new user")
       return done(null, newUser);
     } else {
-      if(banned.includes(profile.id)) return done("You are banned", null)
+      if(checkBan(profile.id)) return done("You are banned", null)
       const newUser = await DiscordUser.create({
         discordId: profile.id,
         username: profile.username,
